revert back to grid, reference too much of issue without hooks
